rename errors/builder to errors/errors
